pyproject.toml
[project]
name = "persona-extraction"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    # Core ML
    "torch>=2.0",
    "transformers>=4.40",
    "peft>=0.10",
    "datasets>=2.18",
    "accelerate>=0.28",
    # APIs
    "anthropic>=0.25",
    "openai>=1.0",
    "httpx>=0.27",
    # Utils
    "pyyaml>=6.0",
    "click>=8.0",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.3"]

[project.scripts]
persona = "src.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
